在 trap_handler 开头加 SBI 跟踪（只针对 pid=2 的非 syscall trap）：
```sh
cd /Users/wangrunji/Codes/iJiege2/kernel && cargo build --release --target riscv64gc-unknown-none-elf 2>&1 | tail -3
```

```sh
cargo build --release --target riscv64gc-unknown-none-elf 2>&1 | grep "error\[" -A5
```

```sh
cargo build --release --target riscv64gc-unknown-none-elf 2>&1 | tail -3 && echo BUILD_OK
```

+++ kernel/src/arch/riscv64/trap/mod.rs
@@ -106,7 +106,7 @@ pub extern "C" fn trap_handler(ctx: &mut TrapContext) {
                         p(if d < 10 { b'0' + d } else { b'a' + d - 10 });
                     }
                 }
-                ps("[2]trap cause="); ph(scause.bits()); ps(" sepc="); ph(ctx.sepc as u64); ps(" stval="); ph(stval as u64); ps("\n");
+                ps("[2]trap cause="); ph(scause.bits() as u64); ps(" sepc="); ph(ctx.sepc as u64); ps(" stval="); ph(stval as u64); ps("\n");
             }
             // 检测 clone（220 的 syscall），但 scause=8 时才是 syscall
         }
